mcp_server/requirements.txt
pandas
numpy
scikit-learn
joblib
requests
fastmcp==2.12.4
google-cloud-storage
